pkg/reconciler: AppStatus doesn't display stale Source
Previously, when a user would try to update an app by pusing again, the
build would seemingly finish instantly. However, the app wasn't actually
updated or readay yet. What was happening was the AppStatus was
returning a stale Source and Build information which was being picked up
by the CLI.

This CL ensures the AppStatus is properly marked as Unknown (instead of
Ready) for app updates.

fixes #372